build: debug file paths
I noticed some files are sending fully qualified paths and not relative
paths to the debug function, which of course could not then match them.

This is an attempt to handle it.  However, it is failing, or rather it
partially works, but causes CORS errors that prevent the layout from
working correctly.