build: custom tool chain check point
remove some vite cruft from server.ts

automate the sidebar generation more fully
this needed to be done in two places,
for dev mode I need it done with the html
for production mode, I need it done *before* the server

I cannot have .ts pages in the Pages directory with this setup.
or rather I could, but it isn't clean. the Right Way is to set up the
route in Hono which means it will be in the server directory.

remove the tsconfig.bin that isn't being used

remove the SidebarSection.spectrum that isn't being used.

fix some include statements that shouldn't have extensions.