build: custom tool chain check point
replaced yaml with  js-yaml - it doesn't use node:process
node:process is incompatible with lambda,
I had to convert some of my own uses of it.
I had to switch to a more robust usage of esbuild
with config file
I switched to cosmiconf from dotenv for configuration
that apparently does a dynamic load of the fs module
and also causes problems. I will have to reeval further