build: check point
converting greenwood to pre-generate much of my content

With the fully custom hono based tool chain I had something I could deploy,
but not run - there were too many places I made choices that ended up
being difficult to run on a lambda/s3 platform in ways that
are not obvious in hono's development mode.

With the plugin based greenwood setup, I had something that was easy to
develop against, but very fragile to build, I was constantly hitting
limitations in my understanding of Greenwood's plugin lifecycle,
typescript's capabilities, limitations in how to deploy a Greenwood
project, AND limitations in lambda.  way too many moving parts.

This approach is trying to reduce the number of moving parts.
I realized that while this is not true of all of my projects,
this project is ENTIRELY amenable to being statically rendered
in advance so long as I render with something that can handle the
custom content bits.

With my new (fragile) experience with rehype transforms I think
I can replace custom elements (except SVG) with transforms and
statically generate the entire page into something that Greenwood
just has to drop into a layout, or potentially even just has to
help me bundle with the appropriate javascript libraries and css
libraries for deployment.

This should drastically reduce the build/bundle risk of
running on lambda, while retaining the "please someone else
understand this bundling thing for me" that Greenwood promises.