build: check point
this is as close as I can come with pre-building into markdown.
I will need a plugin or component to take the place holder and
transform it into html, as the img tag did not work with an svg file,
or rather, it did, but it was incredibly badly formatted because my
css depends on the divs around it.

I could fix that, I may go back and do so, but for now lets explore
this hybrid approach.