build: check point
massively simplifed the update-kv-store.ts, it now runs
but has some hard coded values that will not stay true.
I need to figure out how to properly extract the
current arn for the key store from the stack
then update this to call the describe on it and dynamically get
the first etag.